requirements.txt
line-bot-sdk
flask
numpy
scipy
scikit-learn
jieba
Flask-APScheduler